# web/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "APIREST sobre la BD biblioapp en MySQL Server",
    "version": "1.0.0",
    "title": "BiblioApp API"
  },
  "host": "localhost:8081",
  "basePath": "/biblioapp",
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/auth/logout": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "El usuario cierra sesión.",
        "operationId": "postLogout",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Acción realizada con éxito"
          },
          "204": {
            "description": "No hay datos que mostrar"
          },
          "400": {
            "description": "Error al procesar la petición"
          }
        }
      }
    },
    "/auth/destroysession": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "El usuario destruye la sesión.",
        "operationId": "postDestroy",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Acción realizada con éxito"
          },
          "204": {
            "description": "No hay datos que mostrar"
          },
          "400": {
            "description": "Error al procesar la petición"
          }
        }
      }
    },
...
  },
...
}